replaced custom delete alert component with neeto-UI Alert component
